backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "parsec"
version = "0.1.0"
description = "Document OCR pipeline — PaddleOCR engine with searchable PDF output"
requires-python = ">=3.10"
dependencies = [
    "paddlepaddle>=3.0.0",
    "paddleocr>=3.0.0",
    "ocrmypdf>=17.0.0",
    "jiwer>=3.0.0",
    "pillow>=10.0.0",
    "pdfminer.six",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["parsec*"]
